% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationautoscaling_package.R
\docType{package}
\name{paws.applicationautoscaling-package}
\alias{paws.applicationautoscaling}
\alias{paws.applicationautoscaling-package}
\title{paws.applicationautoscaling: Application Auto Scaling}
\description{
- Amazon ECS services. For more information,
see [Service Auto
Scaling](http://docs.aws.amazon.com/AmazonECS/latest/developerguide/service-auto-scaling.html)
in the *Amazon Elastic Container Service Developer Guide*.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
