% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.applicationautoscaling_package.R
\docType{package}
\name{paws.applicationautoscaling-package}
\alias{paws.applicationautoscaling}
\alias{paws.applicationautoscaling-package}
\title{paws.applicationautoscaling: Application Auto Scaling}
\description{
With Application Auto Scaling, you can configure automatic
scaling for your scalable resources.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
